refactor(cloud-client): introduce records for CreateCluster and CreateZeebeClient response
